CmdGen -- fix input so RANDOM works SELLA --  and Sella_Dev -- work toward making 2 or 3 zeros generate the correct transformations

+++ LRL_Tools/LRL_Vector3.h
@@ -394,7 +394,7 @@ Matrix_3x3(const std::string& s) {
 
 std::vector<double> GetVector() const {
    std::vector<double> v(9);
-   for (size_t i = 0; i < 9; ++i) v[i] = (*this)[i];
+   for (int i = 0; i < 9; ++i) v[i] = (*this)[i];
    return v;
 }
